refactor(goods): 상품 상세 컴포넌트를 FSD 구조에 맞춰 ui 및 widgets으로 분리
