+++ src/features/goods/ui/GoodImage/index.tsx
@@ -1,37 +1,38 @@
-import { useRef, useState } from "react";
 import Image from "next/image";
+import { RefObject } from "react";
 import { Swiper as SwiperType } from "swiper";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Pagination } from "swiper/modules";
-import { GoodDetailImage } from "@/entities/goods/goods.types";
 import { IMAGE_URL } from "@/shared/constants";
+
 import "swiper/css";
 import "swiper/css/pagination";
 import * as s from "./style.css";
+import { GoodDetailImage } from "@/entities/goods/goods.types";
 
 type Props = {
   images: GoodDetailImage[];
+  activeIndex: number;
+  goToSlide: (index: number) => void;
+  onSlideChange: (index: number) => void;
+  swiperRef: RefObject<SwiperType | null>;
 };
 
-export const ThumbnailSwiper = ({ images }: Props) => {
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  const swiperRef = useRef<SwiperType | null>(null);
-
-  const goToSlide = (index: number) => {
-    if (swiperRef.current) {
-      swiperRef.current.slideTo(index);
-    }
-  };
-
+export const GoodImage = ({
+  images,
+  activeIndex,
+  goToSlide,
+  onSlideChange,
+  swiperRef,
+}: Props) => {
   return (
     <div className={s.container}>
       <div className={s.imagesWrapper}>
-        {images?.map((image, idx) => (
+        {images.map((image, idx) => (
           <Image
             key={idx}
             className={s.image({ isActive: activeIndex === idx })}
-            src={`${IMAGE_URL}/${image?.imageUrl}`}
+            src={`${IMAGE_URL}/${image.imageUrl}`}
             alt={`썸네일 이미지 ${idx + 1}`}
             width={56}
             height={56}
@@ -44,19 +45,17 @@ export const ThumbnailSwiper = ({ images }: Props) => {
       <Swiper
         onSwiper={(swiper) => (swiperRef.current = swiper)}
         className={s.swiperWrapper}
-        pagination={{
-          dynamicBullets: true,
-        }}
+        pagination={{ dynamicBullets: true }}
         modules={[Pagination]}
         slidesPerView={1}
         initialSlide={0}
-        onSlideChange={(image) => setActiveIndex(image.activeIndex)}
+        onSlideChange={(swiper) => onSlideChange(swiper.activeIndex)}
       >
-        {images?.map((image, idx) => (
+        {images.map((image, idx) => (
           <SwiperSlide key={idx} className={s.swiperSlide}>
             <Image
               className={s.swiperImage}
-              src={`${IMAGE_URL}/${image?.imageUrl}`}
+              src={`${IMAGE_URL}/${image.imageUrl}`}
               alt={`썸네일 이미지 ${idx + 1}`}
               fill
               sizes="(max-width: 767px) 100vw, 50vw"
