+++ src/features/goods/ui/GoodImage/style.css.ts
