+++ src/features/goods/ui/OptionSelectorItem/style.css.ts
