+++ src/features/goods/ui/OptionSelectorList/style.css.ts
@@ -1,0 +1,10 @@
+import { style } from "@vanilla-extract/css";
+import { f } from "@/shared/styles/functions";
+
+export const wrapper = style([
+  f.flex,
+  f.directionColumn,
+  {
+    gap: "10px",
+  },
+]);
