+++ src/features/goods/ui/SelectedOptionItem/style.css.ts
