+++ src/features/goods/ui/Star/index.tsx
