+++ src/features/goods/ui/TotalPriceInfo/index.tsx
@@ -1,0 +1,15 @@
+import { formatNumberWithCommas } from "@/shared/utils/format/number";
+import * as s from "./style.css";
+
+type Props = {
+  totalPrice: number;
+};
+
+export const TotalPriceInfo = ({ totalPrice }: Props) => {
+  return (
+    <p className={s.priceWrapper}>
+      <span className={s.text}>주문금액</span>
+      <span className={s.price}>{formatNumberWithCommas(totalPrice)}원</span>
+    </p>
+  );
+};
