+++ src/widgets/goods/GoodOptionsSection/style.css.ts
@@ -1,0 +1,4 @@
+import { style } from "@vanilla-extract/css";
+import { f } from "@/shared/styles/functions";
+
+export const container = style([f.wFull]);
